chore: extract cross-suite helpers into src/helpers.ts
Part of REAL-1037.
